Fix all the existing unit tests

+++ app/src/main/java/info/palamarchuk/api/cooking/RecipeEndpoint.java
@@ -1,8 +1,10 @@
 package info.palamarchuk.api.cooking;
 
+import info.palamarchuk.api.cooking.entity.Ingredient;
 import info.palamarchuk.api.cooking.entity.Recipe;
 import info.palamarchuk.api.cooking.entity.RecipeInfo;
 import info.palamarchuk.api.cooking.entity.RecipeIngredient;
+import info.palamarchuk.api.cooking.util.CurrentUrlService;
 import info.palamarchuk.api.cooking.validation.RecipeAddValidator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -11,6 +13,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
+import java.net.URI;
 import java.util.List;
 
 /**
@@ -27,20 +30,14 @@ public class RecipeEndpoint {
         this.service = service;
     }
 
-    @GetMapping(value = "", produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
-    public ResponseEntity<ResponseData<List<Recipe>>> getRecipes() {
-        return new ResponseData<>(service.getAll()).export();
-    }
-
     @GetMapping(value = "/{id}", produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
-    public ResponseEntity<ResponseData<Recipe>> getRecipe(@PathVariable("id") long id) {
+    public ResponseEntity<ResponseData<Recipe>> get(@PathVariable("id") long id) {
         return new ResponseData<>(service.getById(id)).export();
     }
 
-    @GetMapping(value = "/{id}/ingredients", produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
-    public ResponseEntity<ResponseData<List<RecipeIngredient>>> getIngredients(@PathVariable("id") long id) {
-        Recipe recipe = service.getById(id);
-        return new ResponseData<>(recipe.getIngredients()).export();
+    @GetMapping(value = "", produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
+    public ResponseEntity<ResponseData<List<Recipe>>> getAll() {
+        return new ResponseData<>(service.getAll()).export();
     }
 
     @GetMapping(value = "/{id}/infos", produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
@@ -49,40 +46,51 @@ public class RecipeEndpoint {
         return new ResponseData<>(recipe.getInfos()).export();
     }
 
+    @GetMapping(value = "/{id}/ingredients", produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
+    public ResponseEntity<ResponseData<List<RecipeIngredient>>> getIngredients(@PathVariable("id") long id) {
+        Recipe recipe = service.getById(id);
+        return new ResponseData<>(recipe.getIngredients()).export();
+    }
+
     @PostMapping(value = "", consumes = MediaType.APPLICATION_JSON_UTF8_VALUE)
-    public ResponseEntity<ResponseData<Recipe>> addRecipe(@RequestBody Recipe candidate, BindingResult result) {
+    public ResponseEntity<ResponseData<Recipe>> addRecipe(@RequestBody Recipe candidate, @Autowired CurrentUrlService currentUrlService, BindingResult result) {
         new RecipeAddValidator(service).validate(candidate, result);
         if (result.hasErrors()) {
             return new ErrorResponseData(result.getAllErrors()).export(HttpStatus.UNPROCESSABLE_ENTITY);
         }
 
         service.add(candidate);
-        return new ResponseData<>(candidate).export();
+        URI url = currentUrlService.getUrl(candidate.getId());
+        return new ResponseData<>(candidate).exportCreated(url);
     }
 
     @PatchMapping(value = "/{id}", consumes = MediaType.APPLICATION_JSON_UTF8_VALUE)
-    public ResponseEntity<ResponseData<Recipe>> updateRecipe(@PathVariable("id") long id, @RequestBody Recipe candidate) {
+    public ResponseEntity<ResponseData<Recipe>> updateRecipe(@PathVariable("id") long id, @RequestBody Recipe patch) {
         Recipe current = service.getById(id);
         if (current == null) {
             return ResponseEntity.notFound().build(); // @todo Provide additional information
         }
 
-        if (candidate.getCookTime() != null) {
-            current.setCookTime(candidate.getCookTime());
+        if (patch.getCookTime() != null) {
+            current.setCookTime(patch.getCookTime());
         }
-        if (candidate.getPrecookTime() != null) {
-            current.setCookTime(candidate.getPrecookTime());
+        if (patch.getPrecookTime() != null) {
+            current.setCookTime(patch.getPrecookTime());
         }
         if (current.getName() != null) {
-            current.setName(candidate.getName());
+            current.setName(patch.getName());
         }
 
-        service.update(candidate);
-        return new ResponseData<>(candidate).export();
+        service.update(current);
+        return new ResponseData<>(current).export();
     }
 
     @DeleteMapping(value = "/{id}")
     public ResponseEntity deleteRecipe(@PathVariable("id") long id) {
+        Recipe current = service.getById(id);
+        if (current == null) {
+            return ResponseEntity.notFound().build();
+        }
         service.deleteById(id);
         return ResponseEntity.noContent().build();
     }
